New test-suite for control-thread testing. Divided peer thread so that it can be mocked.
