+++ functorrent.cabal
@@ -23,9 +23,13 @@ library
                        FuncTorrent.Network
                        FuncTorrent.Peer,
                        FuncTorrent.Tracker
+                       FuncTorrent.ControlThread
+                       FuncTorrent.PeerThread
+                       FuncTorrent.PeerThreadMain
+                       FuncTorrent.PeerThreadData
 
   other-modules:       FuncTorrent.Utils
-  other-extensions:    OverloadedStrings
+  other-extensions:    OverloadedStrings, CPP
   hs-source-dirs:      src
   ghc-options:         -Wall -fwarn-incomplete-patterns -fno-warn-orphans
   default-language:    Haskell2010
@@ -42,7 +46,8 @@ library
                        parsec,
                        QuickCheck,
                        tasty,
-                       tasty-hunit
+                       tasty-hunit,
+                       lens
 
 executable functorrent
   main-is:             Main.hs
@@ -61,7 +66,8 @@ executable functorrent
                        QuickCheck,
                        network,
                        network-uri,
-                       parsec
+                       parsec,
+                       lens
 
 test-suite functorrent-test
   type:              exitcode-stdio-1.0
@@ -81,3 +87,31 @@ test-suite functorrent-test
                      QuickCheck,
                      tasty-quickcheck,
                      test-framework-quickcheck2
+
+test-suite control-thread-tests
+  type:              exitcode-stdio-1.0
+  default-language:  Haskell2010
+  hs-source-dirs:    test,src
+  main-is:           ControlThreadTestsMain.hs
+  other-modules:     PeerThreadMock
+                     ControlThreadTests
+                     FuncTorrent.ControlThread
+  ghc-options:       -Wall
+  other-extensions:  CPP
+  cpp-options:       -DPEER_THREAD_MOCKED
+  build-depends:     base,
+                     base16-bytestring,
+                     binary,
+                     bytestring,
+                     containers,
+                     cryptohash,
+                     directory,
+                     HTTP,
+                     network,
+                     network-uri,
+                     parsec,
+                     QuickCheck,
+                     tasty,
+                     tasty-hunit,
+                     functorrent,
+                     lens
