+++ test/ControlThreadTestsMain.hs
@@ -1,0 +1,7 @@
+{-# LANGUAGE OverloadedStrings #-}
+module Main where
+
+import ControlThreadTests
+
+-- Merge this with the functorrent-test once this becomes stable
+main = doTests
